Refactor Groups and GroupSingle and  resolve group leader photo
