+++ src/group-single/getGroup.js
@@ -1,0 +1,37 @@
+import gql from 'graphql-tag';
+
+export const GROUP_FRAGMENT = gql`
+  fragment groupFragment on Group {
+    id
+    name
+    title
+    summary
+    members {
+      id
+      firstName
+      photo {
+        uri
+      }
+    }
+    leaders {
+      id
+      firstName
+      photo {
+        uri
+      }
+    }
+    schedule
+  }
+`;
+
+export default gql`
+  query getGroup($itemId: ID!) {
+    node(id: $itemId) {
+      __typename
+      ... on Group {
+        ...groupFragment
+      }
+    }
+  }
+  ${GROUP_FRAGMENT}
+`;
